+++ src/index.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const app = require('express')();
 const session = require('express-session');
 const port = 8000;
@@ -5,6 +7,7 @@ const login = require('./routes/login');
 const logout = require('./routes/logout');
 const path = require('path');
 const bodyParser = require('body-parser');
+const proxy = require('./routes/proxy');
 
 app.use(session({
     secret: 'losspolsss',
@@ -17,9 +20,11 @@ app.use(bodyParser.urlencoded({extended : true}));
 app.use('/login', login);
 app.use('/logout', logout);
 
+app.use('/', proxy);
+
 app.get('/', (request, response) => {
     if(request.session.user) {
-        response.send({session: request.session.user});
+        response.redirect('/chat');
     } else {
         response.sendFile(path.join(__dirname + '/views/login.html'));
     }